+++ addons/TBMod_zusatzStuff/eos/CfgFunctions.hpp
@@ -1,13 +1,28 @@
 ﻿/*
     Author: Willi "shukari" Graff
 */
-class TBMod_zusatzStuff_eos
+class TBMod_zusatzStuff_LV
 {
-    tag = "TB_EOS";
+    tag = "TB_LV";
+    
+    class modules
+    {
+        file = "\TBMod_zusatzStuff\LV";
+        class militarize {};
+        class swimPatrol {};
+        class moduleMilitarize {};
+    };
     
     class functions
     {
-        file = "\TBMod_zusatzStuff\eos\functions";
-        //class UpDown {};
+        file = "\TBMod_zusatzStuff\LV\functions";
+        class classnames {};
+        class validateClassArrays {};
+        class ACskills {};
+        class fullAirVehicle {};
+        class fullLandVehicle {};
+        class fullWaterVehicle {};
+        class ACAIkilled {};
+        class getPlayers {};
     };
 };
