+++ addons/TBMod_zusatzStuff/eos/core/fn_b_Launch.sqf
