--- addons/TBMod_zusatzStuff/eos/functions/EOS_Markers.sqf
@@ -1,5 +1,0 @@
-_eosMarkers=server getvariable "EOSmarkers";
-
-{_x setMarkerAlpha (MarkerAlpha _x);
-_x setMarkercolor (getMarkercolor _x);
-}foreach _eosMarkers;
