+++ addons/TBMod_zusatzStuff/eos/functions/fn_shk_buildingpos.sqf
